Create Smart Boiler Control
The acquisition of data of various parameters related to the boiler system viz. surrounding temperature, humidity, water fluid pressure can be done from the sensor using an intelligent device microcontroller Arduino Uno. The acquired data was processed to monitor the performance of the boiler on the smart device via wireless Wi-Fi connectivity. The boiler can be switched ON/ OFF from the smart device remotely without physical interaction with the boiler. The communication between the microcontroller board and a smart device is possible by virtue of IoSD.  
The system aims to enhance the user experience for boiler control and enables the user to control and monitor the performance of the boiler. On a webpage, user can also see the surrounding temperature in order to judge the power consumption of the boiler. More on the system is capable of taking smart decisions in case danger to prevent the possible damage.